% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20007R.R
\name{mardias}
\alias{mardias}
\title{A simplified Mardia multivariate skewness and kurtosis calculator}
\usage{
mardias(x, na.rm = TRUE)
}
\arguments{
\item{x}{A data matrix}

\item{na.rm}{An indication of the missing data, the default value is True}
}
\value{
Mardia multivariate skewness and kurtosis
}
\description{
A simplified Mardia multivariate skewness and kurtosis calculator
}
\examples{
\dontrun{
x <- mvrnorm(1000,c(1,2,3),matrix(c(4,0.7,0.3,0.7,2,1.4,0.3,1.4,5),nrow=3,ncol=3))
mardias(x)
}
}
