% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20007R.R
\name{mstore}
\alias{mstore}
\title{Moment calculator of multivariate normal distribution}
\usage{
mstore(p, sig)
}
\arguments{
\item{p}{Dimension}

\item{sig}{p*p multivariate normal covariance matrix}
}
\value{
Moments flattened into an vector
}
\description{
Moment calculator of multivariate normal distribution
}
\examples{
\dontrun{
p <- 2
sig <- matrix(c(2,0.7,0.7,1),2,2)
A <- mstore(p,sig)
matrix(A,13,13)
}
}
