% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwC}
\alias{rwC}
\title{A random walk Metropolis sampler using Rcpp}
\usage{
rwC(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard deviation used for the proposal distribution}

\item{x0}{the initial value}

\item{N}{the number of samples}
}
\value{
a random sample of size N
}
\description{
A random walk Metropolis sampler using Rcpp
}
\examples{
\dontrun{
sigma <- 2
x0 <- 5
N <- 2000
rnC <- rwC(sigma, x0, N)
}
}
